Project ALKES/requirements.txt
# pip install -r requirements.txt

PyQt6
openpyxl
pyinstaller
KlAkOAPI-13.tar.gz